feat: refactor naming
