+++ packages/cli-tools/src/services/upload/upload.batch.services.ts
@@ -1,15 +1,13 @@
-import {isNullish, nonNullish, notEmptyString} from '@dfinity/utils';
-import {readFile} from 'node:fs/promises';
-import {basename} from 'node:path';
+import {isNullish, nonNullish} from '@dfinity/utils';
 import {
   type COLLECTION_CDN_RELEASES,
   type COLLECTION_DAPP,
   UPLOAD_BATCH_SIZE
-} from '../constants/deploy.constants';
-import type {FileAndPaths, FileDetails, UploadFiles, UploadFileStorage} from '../types/deploy';
+} from '../../constants/deploy.constants';
+import type {FileAndPaths, UploadFiles} from '../../types/deploy';
+import {prepareFileForUpload} from './_upload.services';
 
-// TODO: rename
-export const uploadFilesGrouped = async ({
+export const uploadFilesWithBatch = async ({
   files: sourceFiles,
   uploadFiles,
   sourceAbsolutePath,
@@ -96,30 +94,3 @@ const uploadFilesToStorage = async ({
     files: filesToUpload
   });
 };
-
-// TODO: duplicated code
-const prepareFileForUpload = async ({
-  file,
-  fullPath,
-  collection,
-  filePath,
-  token,
-  description
-}: {
-  file: FileDetails;
-  filePath: string;
-} & Pick<
-  UploadFileStorage,
-  'fullPath' | 'collection' | 'token' | 'description'
->): Promise<UploadFileStorage> => ({
-  filename: basename(filePath),
-  fullPath,
-  data: new Blob([await readFile(file.file)]),
-  collection,
-  headers: [
-    ...(file.mime === undefined ? [] : ([['Content-Type', file.mime]] as Array<[string, string]>))
-  ],
-  encoding: file.encoding,
-  ...(nonNullish(token) && notEmptyString(token) && {token}),
-  ...(nonNullish(description) && notEmptyString(description) && {description})
-});
